refactor: add file to ignore non-breaking warnings and error logs
