+++ src/services/ignoreWarnings.js
@@ -1,0 +1,13 @@
+if (__DEV__) {
+  const ignoreWarns = ["VirtualizedLists should never be nested inside plain ScrollViews"];
+
+  const errorWarn = global.console.error;
+  global.console.error = (...arg) => {
+    for (const error of ignoreWarns) {
+      if (arg[0].startsWith(error)) {
+        return;
+      }
+    }
+    errorWarn(...arg);
+  };
+}
